Rename auth package to authenticator
