+++ backend/app/adapter/graphql/graphql.go
@@ -2,7 +2,8 @@ package graphql
 
 import (
 	"github.com/short-d/short/app/adapter/graphql/resolver"
-	"github.com/short-d/short/app/usecase/auth"
+
+	"github.com/short-d/short/app/usecase/authenticator"
 	"github.com/short-d/short/app/usecase/changelog"
 	"github.com/short-d/short/app/usecase/requester"
 	"github.com/short-d/short/app/usecase/url"
@@ -35,7 +36,7 @@ func NewShort(
 	urlCreator url.Creator,
 	changeLog changelog.ChangeLog,
 	requesterVerifier requester.Verifier,
-	authenticator auth.Authenticator,
+	authenticator authenticator.Authenticator,
 ) Short {
 	r := resolver.NewResolver(
 		logger,
