+++ backend/app/adapter/graphql/graphql_test.go
@@ -8,7 +8,8 @@ import (
 
 	"github.com/short-d/app/mdtest"
 	"github.com/short-d/short/app/adapter/db"
-	"github.com/short-d/short/app/usecase/auth"
+
+	"github.com/short-d/short/app/usecase/authenticator"
 	"github.com/short-d/short/app/usecase/changelog"
 	"github.com/short-d/short/app/usecase/keygen"
 	"github.com/short-d/short/app/usecase/requester"
@@ -41,7 +42,7 @@ func TestGraphQlAPI(t *testing.T) {
 
 	s := service.NewReCaptchaFake(service.VerifyResponse{})
 	verifier := requester.NewVerifier(s)
-	authenticator := auth.NewAuthenticatorFake(time.Now(), time.Hour)
+	authenticator := authenticator.NewAuthenticatorFake(time.Now(), time.Hour)
 
 	logger := mdtest.NewLoggerFake(mdtest.FakeLoggerArgs{})
 	tracer := mdtest.NewTracerFake()
