+++ backend/app/adapter/graphql/resolver/authquery.go
@@ -5,7 +5,8 @@ import (
 
 	"github.com/short-d/short/app/adapter/graphql/scalar"
 	"github.com/short-d/short/app/entity"
-	"github.com/short-d/short/app/usecase/auth"
+
+	"github.com/short-d/short/app/usecase/authenticator"
 	"github.com/short-d/short/app/usecase/changelog"
 	"github.com/short-d/short/app/usecase/url"
 )
@@ -14,7 +15,7 @@ import (
 // on the identify of the user
 type AuthQuery struct {
 	authToken     *string
-	authenticator auth.Authenticator
+	authenticator authenticator.Authenticator
 	changeLog     changelog.ChangeLog
 	urlRetriever  url.Retriever
 }
@@ -73,7 +74,7 @@ func (v AuthQuery) URLs() ([]URL, error) {
 
 func newAuthQuery(
 	authToken *string,
-	authenticator auth.Authenticator,
+	authenticator authenticator.Authenticator,
 	changeLog changelog.ChangeLog,
 	urlRetriever url.Retriever,
 ) AuthQuery {
