+++ backend/app/adapter/graphql/resolver/viewer.go
@@ -4,10 +4,10 @@ import (
 	"errors"
 
 	"github.com/short-d/short/app/entity"
-	"github.com/short-d/short/app/usecase/auth"
+	"github.com/short-d/short/app/usecase/authenticator"
 )
 
-func viewer(authToken *string, authenticator auth.Authenticator) (entity.User, error) {
+func viewer(authToken *string, authenticator authenticator.Authenticator) (entity.User, error) {
 	if authToken == nil {
 		return entity.User{}, errors.New("auth token can't be empty")
 	}
