+++ backend/app/usecase/authenticator/authenticator_fake.go
@@ -1,4 +1,4 @@
-package auth
+package authenticator
 
 import (
 	"time"
