+++ backend/app/usecase/authenticator/authenticator_test.go
@@ -1,6 +1,6 @@
 // +build !integration all
 
-package auth
+package authenticator
 
 import (
 	"testing"
