+++ backend/app/usecase/authenticator/payloadtest/factory_stub.go
@@ -3,7 +3,7 @@ package payloadtest
 import (
 	"github.com/short-d/app/fw"
 	"github.com/short-d/short/app/entity"
-	"github.com/short-d/short/app/usecase/auth/payload"
+	"github.com/short-d/short/app/usecase/authenticator/payload"
 )
 
 var _ payload.Factory = (*FactoryStub)(nil)
