+++ backend/dep/provider/routes.go
@@ -8,7 +8,8 @@ import (
 	"github.com/short-d/short/app/adapter/request"
 	"github.com/short-d/short/app/adapter/routing"
 	"github.com/short-d/short/app/usecase/account"
-	"github.com/short-d/short/app/usecase/auth"
+	"github.com/short-d/short/app/usecase/authenticator"
+
 	"github.com/short-d/short/app/usecase/feature"
 	"github.com/short-d/short/app/usecase/url"
 )
@@ -26,7 +27,7 @@ func NewShortRoutes(
 	facebookAPI facebook.API,
 	googleAPI google.API,
 	featureDecisionFactory feature.DecisionFactory,
-	authenticator auth.Authenticator,
+	authenticator authenticator.Authenticator,
 	accountProvider account.Provider,
 ) []fw.Route {
 	return routing.NewShort(
